Drop GenericCamera
